[Fixed] Protypes fixed

+++ 0x08-palindrome_integer/palindrome.h
@@ -1,7 +1,7 @@
 #ifndef PALINDROME_H
 #define PALINDROME_H
 
-unsigned long is_palindrome(unsigned long n);
-int reversed(unsigned long n);
+int is_palindrome(unsigned long n);
+unsigned long reversed(unsigned long n);
 
 #endif
